サンプルファイル削除,エラーメッセージ変更
