+++ src/index.tsx
@@ -3,15 +3,13 @@ import ReactDOM from 'react-dom';
 import CssBaseline from '@mui/material/CssBaseline';
 import { ThemeProvider } from '@mui/material/styles';
 import App from './App';
-import ValidationSample from './ValidationSample';
-import ValidationSampleMui from './ValidationSampleMui';
 import theme from './theme';
 
 ReactDOM.render(
   <ThemeProvider theme={theme}>
     {/* CssBaseline kickstart an elegant, consistent, and simple baseline to build upon. */}
     <CssBaseline />
-    <ValidationSampleMui />
+    <App />
   </ThemeProvider>,
   document.querySelector('#root')
 );
